# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AsistenteVenta API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CarAttractiveness =
  (typeof CarAttractiveness)[keyof typeof CarAttractiveness];

export const CarAttractiveness = {
  hot: "hot",
  normal: "normal",
  hard: "hard",
} as const;

export type CarStatus = (typeof CarStatus)[keyof typeof CarStatus];

export const CarStatus = {
  open: "open",
  locking: "locking",
  locked: "locked",
  released: "released",
  sold: "sold",
} as const;

export interface Car {
  id: number;
  make: string;
  model: string;
  year: number;
  price: number;
  priceLabel: string;
  attractiveness: CarAttractiveness;
  windowHours: number;
  status: CarStatus;
  /** @nullable */
  imageUrl?: string | null;
  km: number;
  fuel: string;
  transmission: string;
  location: string;
  depositCents: number;
  publishedAt: string;
  availableUntil: string;
  /** @nullable */
  lockedUntil?: string | null;
  /** @nullable */
  releasedAt?: string | null;
  viewersNow: number;
  /** @nullable */
  notes?: string | null;
}

export type CreateCarBodyAttractiveness =
  (typeof CreateCarBodyAttractiveness)[keyof typeof CreateCarBodyAttractiveness];

export const CreateCarBodyAttractiveness = {
  hot: "hot",
  normal: "normal",
  hard: "hard",
} as const;

export interface CreateCarBody {
  make: string;
  model: string;
  year: number;
  price: number;
  attractiveness: CreateCarBodyAttractiveness;
  km: number;
  fuel: string;
  transmission: string;
  location: string;
  /** @nullable */
  imageUrl?: string | null;
  depositCents: number;
  /** @nullable */
  notes?: string | null;
}

export type UpdateCarBodyAttractiveness =
  (typeof UpdateCarBodyAttractiveness)[keyof typeof UpdateCarBodyAttractiveness];

export const UpdateCarBodyAttractiveness = {
  hot: "hot",
  normal: "normal",
  hard: "hard",
} as const;

export interface UpdateCarBody {
  price?: number;
  attractiveness?: UpdateCarBodyAttractiveness;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface LockCarBody {
  leadId: number;
  depositCents?: number;
}

export type LeadStage = (typeof LeadStage)[keyof typeof LeadStage];

export const LeadStage = {
  new: "new",
  awaiting_deposit: "awaiting_deposit",
  locked: "locked",
  doubting: "doubting",
  released: "released",
  won: "won",
  lost: "lost",
} as const;

export interface Lead {
  id: number;
  name: string;
  phone: string;
  carId: number;
  stage: LeadStage;
  depositPaid: boolean;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  lastMessageAt?: string | null;
  /** @nullable */
  lastMessagePreview?: string | null;
  unreadCount: number;
  avatarColor: string;
}

export type LeadWithCar = Lead & {
  car: Car;
};

export type MessageDirection =
  (typeof MessageDirection)[keyof typeof MessageDirection];

export const MessageDirection = {
  incoming: "incoming",
  outgoing: "outgoing",
  system: "system",
} as const;

export interface Message {
  id: number;
  leadId: number;
  direction: MessageDirection;
  content: string;
  createdAt: string;
  aiGenerated: boolean;
  /** @nullable */
  intent?: string | null;
}

export type LeadDetail = LeadWithCar & {
  messages: Message[];
};

export interface CreateLeadBody {
  name: string;
  phone: string;
  carId: number;
}

export type UpdateLeadBodyStage =
  (typeof UpdateLeadBodyStage)[keyof typeof UpdateLeadBodyStage];

export const UpdateLeadBodyStage = {
  new: "new",
  awaiting_deposit: "awaiting_deposit",
  locked: "locked",
  doubting: "doubting",
  released: "released",
  won: "won",
  lost: "lost",
} as const;

export interface UpdateLeadBody {
  stage?: UpdateLeadBodyStage;
  depositPaid?: boolean;
  name?: string;
  phone?: string;
}

export interface SendMessageBody {
  content: string;
  aiGenerated?: boolean;
  /** @nullable */
  intent?: string | null;
}

export interface SimulateIncomingBody {
  content: string;
}

export type DraftReplyBodyIntent =
  (typeof DraftReplyBodyIntent)[keyof typeof DraftReplyBodyIntent];

export const DraftReplyBodyIntent = {
  first_response: "first_response",
  ask_deposit: "ask_deposit",
  confirm_lock: "confirm_lock",
  handle_doubt: "handle_doubt",
  post_release: "post_release",
  custom: "custom",
} as const;

export interface DraftReplyBody {
  intent: DraftReplyBodyIntent;
  /** @nullable */
  instructions?: string | null;
}

export interface DraftReply {
  intent: string;
  content: string;
  rationale: string;
}

export interface StageCount {
  stage: string;
  count: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface DashboardSummary {
  openCars: number;
  lockedCars: number;
  releasedCars: number;
  soldCars: number;
  activeLeads: number;
  awaitingDeposit: number;
  wonLast7d: number;
  depositValueCents: number;
  avgResponseMinutes: number;
  conversionRate: number;
  leadsByStage: StageCount[];
  carsByStatus: StatusCount[];
}

export type ActivityEventKind =
  (typeof ActivityEventKind)[keyof typeof ActivityEventKind];

export const ActivityEventKind = {
  lock: "lock",
  release: "release",
  sale: "sale",
  message: "message",
  draft: "draft",
  new_lead: "new_lead",
} as const;

export interface ActivityEvent {
  id: number;
  kind: ActivityEventKind;
  text: string;
  /** @nullable */
  leadName?: string | null;
  /** @nullable */
  carLabel?: string | null;
  at: string;
}

export type ListCarsParams = {
  status?: ListCarsStatus;
};

export type ListCarsStatus =
  (typeof ListCarsStatus)[keyof typeof ListCarsStatus];

export const ListCarsStatus = {
  open: "open",
  locking: "locking",
  locked: "locked",
  released: "released",
  sold: "sold",
} as const;

export type ListLeadsParams = {
  stage?: ListLeadsStage;
};

export type ListLeadsStage =
  (typeof ListLeadsStage)[keyof typeof ListLeadsStage];

export const ListLeadsStage = {
  new: "new",
  awaiting_deposit: "awaiting_deposit",
  locked: "locked",
  doubting: "doubting",
  released: "released",
  won: "won",
  lost: "lost",
} as const;