# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AsistenteVenta API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEventKind } from "./activityEventKind";

export interface ActivityEvent {
  id: number;
  kind: ActivityEventKind;
  text: string;
  /** @nullable */
  leadName?: string | null;
  /** @nullable */
  carLabel?: string | null;
  at: Date;
}