# lib/api-zod/src/generated/types/createLeadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AsistenteVenta API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateLeadBody {
  name: string;
  phone: string;
  carId: number;
}