# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AsistenteVenta API
 * OpenAPI spec version: 0.1.0
 */
import type { StageCount } from "./stageCount";
import type { StatusCount } from "./statusCount";

export interface DashboardSummary {
  openCars: number;
  lockedCars: number;
  releasedCars: number;
  soldCars: number;
  activeLeads: number;
  awaitingDeposit: number;
  wonLast7d: number;
  depositValueCents: number;
  avgResponseMinutes: number;
  conversionRate: number;
  leadsByStage: StageCount[];
  carsByStatus: StatusCount[];
}